Extract LimeImportsCollector as a stand-alone class
As a part of the effort of aligning code structure across all
generators, extracted imports collection logic from
LimeGenerator into a stand-alone class LimeImportsCollector.

Also added sorting to the list of imports post-collection,
similar to what most generators already do. Updated smoke
tests affected by this sorting.

See: #809
