+++ gluecodium/src/main/java/com/here/gluecodium/generator/common/ImportsResolver.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright (C) 2016-2021 HERE Europe B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package com.here.gluecodium.generator.common
+
+import com.here.gluecodium.model.lime.LimeNamedElement
+
+/** Interface for resolving imports for the given element, but not any nested declarations. */
+fun interface ImportsResolver<T> {
+    fun resolveElementImports(limeElement: LimeNamedElement): List<T>
+}
