+++ gluecodium/src/test/resources/smoke/escaped_names/output/lime/package/class.lime
@@ -1,16 +1,12 @@
 package `package`
-
 import `package`.`types`.`List`
-import `package`.`types`.`struct`
-import `package`.`types`.`exception`
 import `package`.`types`.`enum`
-
+import `package`.`types`.`exception`
+import `package`.`types`.`struct`
 class `class` : interface {
     constructor `constructor`()
-
     fun `fun`(
         `Double`: `List`
     ): `struct` throws `exception`
-
     property `property`: `enum`
 }
